bundle: Temporarily purge all bundles (stored as git submodules).
Time to start moving the auto.py script out of beta.
I'll start using this script from now on to handle
automating submodule/bundle management.

Before that though, I want to purge all current bundles
from this repository as a sort of intermediate commit so
that I can start with a clean slate.

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>