install: Update setup.sh and vimrc for a warning-less installation.
The current installation throws a warning because we try to
set the colorscheme to vim-one before we've actually
installed it.

So a simple workaround is that I can add the default colorscheme
configuration to the vimrc after I've installed vim-one and all
of the other plugins.

Locally, I'm going to put the colorscheme stuff in local.vim
since I change it up quite frequently based on how bright it
is where I'm sitting.

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>