ftplugin/go: Fix golang forced autocomplete issue.
Long story short, replacing . with .<C-x><C-o> would have
automatically filled in the first valid function for
that position. - which is often not the function we want.

This is pretty annoying.

Instead we replace . with .<C-x><C-o><C-P> which will keep
the omnicomplete menu open but not auto-fill in the first
function (rather <C-x><C-o> fills it in, but then the
following <C-P> removes it, whilst keeps the menu open).

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>