core: Setup 2 vertical lines, one at 80 and one at 100.
Ideally code is only supposed to be 80 chars long, and this
helps a bunch when working with multiple splits. But let's
be honest - that's not always possible. So what I go for
is "Casually try to keep it below 80 but not if it means
screwing up readability. Then try to keep it below 100 with a
bit more effort."