plugged: Remove empty folders, will be created on PlugInstall.
These folders are supposed to be created when we actually install
the plugins. The only things this repo provides offline are static
copies of:
    - script: junegunn/vim-plug
    - colorscheme: arcticicestudio/nord-vim
    - colorscheme: lifepillar/vim-solarized8
All of which are under an MIT License (so it's ok to include static
copies).

It would be nice to provide static copies of the default plugins,
but that would be kinda weird to stuff into this repository (they
would be entire folders instead of just a single file).

Git submodules would be redundant here since PlugInstall gets the
plugins directly from GitHub either way.